refactor: Move SearchSettingsModule to modules dir
